python/requirements.txt
rerun-sdk
pyzmq
numpy
